⚙️ Refactor page components
