--- client/src/pages/Profile.tsx
@@ -1,5 +1,0 @@
-const Profile = () => {
-  return <div>Profile</div>;
-};
-
-export default Profile;
