+++ client/src/pages/SearchPage.tsx
@@ -1,0 +1,5 @@
+const SearchPage = () => {
+  return <div>Search</div>;
+};
+
+export default SearchPage;
